Refactored file layout to allow for multiple build targets
