--- src/BraneEngine.h
@@ -1,11 +1,0 @@
-﻿#pragma once
-
-#include <iostream>
-#include "ECS/core/Entity.h"
-#include "Graphics/graphics.h"
-#include "clock.h"
-
-// Make sure our debug code is only included in debug builds
-#ifndef NDEBUG
-#include <tests.h>
-#endif
