+++ src/CMakeLists.txt
@@ -1,16 +1,34 @@
 #file: src
 
 
-file(GLOB_RECURSE LSOURCES *.cpp *.h)
+include_directories(
+	${CMAKE_CURRENT_SOURCE_DIR}
+	${CMAKE_CURRENT_SOURCE_DIR}/common
+	../include
+)
 
-message("SOURCES: " ${SOURCES})
+# run child cmake lists
+add_subdirectory(common)
 
-add_executable(BraneEngine ${LSOURCES} ${SOURCES} )
-target_include_directories(BraneEngine PRIVATE 
+# create executables
+add_subdirectory(client)
+
+
+set (CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/runtimeServer)
+add_executable(BraneRuntimeServer runtimeServer/main.cpp)
+
+message("LIBS: " ${LIBS})
+target_link_libraries(BraneRuntimeServer PRIVATE ${LIBS})
+target_compile_definitions(BraneRuntimeServer PRIVATE ${COMMON_DEFS} ${ASSET_SERVER_DEFS})
+
+set (CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/assetServer)
+add_executable(BraneAssetServer assetServer/main.cpp)
+target_include_directories(BraneAssetServer PRIVATE 
 	${INCS}
 	${CMAKE_CURRENT_SOURCE_DIR}
 	ECS
 	Tests
 )
 message("LIBS: " ${LIBS})
-target_link_libraries(BraneEngine PRIVATE ${LIBS})
+target_link_libraries(BraneAssetServer PRIVATE ${LIBS})
+target_compile_definitions(BraneAssetServer PRIVATE ${COMMON_DEFS} ${ASSET_SERVER_DEFS})
