+++ src/common/CMakeLists.txt
@@ -1,0 +1,3 @@
+# src/common
+add_subdirectory(utility)
+add_subdirectory(ecs)
