+++ src/common/ecs/core/archetype.h
@@ -5,7 +5,7 @@
 #include <memory>
 #include <functional>
 #include <unordered_map>
-#include "stackAllocate.h"
+#include <common/utility/stackAllocate.h>
 
 typedef uint64_t ArchetypeID;
 class VirtualArchetype;
