+++ src/common/ecs/core/component.h
