+++ src/common/ecs/core/virtualType.h
