+++ src/common/utility/CMakeLists.txt
@@ -1,0 +1,6 @@
+# src/common/utlity
+
+add_library(utility STATIC
+	clock.cpp
+	threadPool.cpp
+)
