+++ src/common/utility/clock.cpp
