+++ src/listFiles.bat
@@ -1,0 +1,1 @@
+python braggingStats.py -n
